HTML external rules working.
Changed XML and HTML classes to *RuleExtractor. They no longer create objects, and the name says what they are for. Using objects was causing problems when the HTML module had to call the external rule extractor contained in the XML class, and it was just the wrong thing for the job and required extra work.